LayoutTests/media/media-source/media-managedmse-airplay-expected.txt


RUN(source = new ManagedMediaSource())
RUN(video.src = URL.createObjectURL(source))
RUN(video.disableRemotePlayback = true)
RUN(sourceOpenAllowed = true)
EVENT(sourceopen)
EVENT(sourceopen)
RUN(sourceOpenAllowed = false)
RUN(video.disableRemotePlayback = false)
RUN(source = new ManagedMediaSource())
RUN(video.removeAttribute("src"))
RUN(sourceElement1 = document.createElement("source"))
RUN(sourceElement1.src = URL.createObjectURL(source))
RUN(video.appendChild(sourceElement1))
RUN(sourceElement2 = document.createElement("source"))
RUN(sourceElement2.src = "http://foo.com/playlist.m3u8")
RUN(sourceElement2.type = "application/vnd.apple.mpegurl")
RUN(video.appendChild(sourceElement2))
RUN(sourceOpenAllowed = true)
EVENT(sourceopen)
END OF TEST
